refactor Card
