+++ components/Card/FindMatePostCard.tsx
@@ -1,4 +1,4 @@
-import Card from './Card';
+import Card from './components';
 
 type Props = {
   thumbnail: string;
