+++ components/Card/components/IconLinks.tsx
@@ -1,19 +1,28 @@
 import { RiKakaoTalkFill } from 'react-icons/ri';
 import { SiDiscord } from 'react-icons/si';
 import styles from './IconLinks.module.scss';
+import cx from '@/styles/cx';
+
+// const classNames = cx(styles)('button', size, color, { outline, disabled });
 
 const DiscordLink = ({ link }: { link: string }) => {
+  const className = cx(styles)('linkContainer', 'discord');
   return (
-    <a className={styles.discordLink} href={link}>
-      <SiDiscord />
+    <a className={className} href={link}>
+      <div>
+        <SiDiscord />
+      </div>
     </a>
   );
 };
 
 const KakaoLink = ({ link }: { link: string }) => {
+  const className = cx(styles)('linkContainer', 'kakao');
   return (
-    <a className={styles.kakaoLink} href={link}>
-      <RiKakaoTalkFill />
+    <a className={className} href={link}>
+      <div>
+        <RiKakaoTalkFill />
+      </div>
     </a>
   );
 };
