+++ components/Card/index.ts
@@ -1,0 +1,3 @@
+import FindMatePostCard from './FindMatePostCard';
+
+export { FindMatePostCard };
